chore: update version to 1.24.3rc0 and refactor module imports; consolidate module constants
